requirements.txt
mysql-connector-python==8.3.0
Pillow==11.2.1
ttkthemes==3.2.2
reportlab==4.4.1
